Implemented the ability to create sub-categories.

+++ trunk/swisscenter/config/config_category.php
@@ -11,22 +11,25 @@
   {
     $cat           = ( isset($_REQUEST["cat"]) ? $_REQUEST["cat"] : '');
     $cat_name      = ( isset($_REQUEST["cat_name"]) ? un_magic_quote($_REQUEST["cat_name"]) : '');
+    $parent_names  = db_toarray('select cat_id VAL, cat_name NAME from categories order by cat_name');
     $download_opts = array( array("VAL"=>'N',"NAME"=> str('DISABLED') )
                           , array("VAL"=>'Y',"NAME"=> str('ENABLED')  ));
 
     if(empty($cat))
     {
       // Get a list of all of the cats from the database and display them
-      $data = db_toarray("SELECT cat_id
-                               , cat_name 'Category'
-                               , (CASE download_info
+      $data = db_toarray("SELECT DISTINCT c.cat_id
+                               , c.cat_name 'Category'
+                               , p.cat_name 'Category_Parent'
+                               , (CASE c.download_info
                                   WHEN 'Y' THEN '".str('ENABLED')."'
                                   WHEN 'N' THEN '".str('DISABLED')."'
                                   ELSE 'Unknown' 
                                   END
                                  ) download_info
-                            FROM categories
-                        ORDER BY Category");
+                            FROM categories c, categories p
+                           WHERE c.parent_id=p.cat_id
+                        ORDER BY p.cat_id,Category");
       
       echo "<h1>".str('CATEGORIES')."</h1>";
       message($del_message);
@@ -34,9 +37,9 @@
       form_hidden('section', 'CATEGORY');
       form_hidden('action', 'MODIFY');
 
-      form_select_table('cat_ids', $data, str('NAME').','.str('CAT_DOWNLOAD')
+      form_select_table('cat_ids', $data, str('NAME').','.str('CAT_PARENT').','.str('CAT_DOWNLOAD')
                        ,array('class'=>'form_select_tab','width'=>'100%'), 'cat_id'
-                       ,array('CATEGORY'=>'','DOWNLOAD_INFO'=>$download_opts)
+                       ,array('CATEGORY'=>'','CATEGORY_PARENT'=>$parent_names,'DOWNLOAD_INFO'=>$download_opts)
                        , $edit_id, 'cats');
       form_submit(str('CAT_DEL_BUTTON'), 1, 'center');
       form_end();
@@ -47,6 +50,7 @@
       form_hidden('section', 'CATEGORY');
       form_hidden('action', 'ADD');
       form_input('cat_name', str('NAME'), 60, 100, $cat_name);
+      form_list_dynamic('parent_id', str('CAT_PARENT'),"select cat_id,cat_name from categories order by cat_name", $_REQUEST['parent_id']);
 
       // Only display the "Download Info" options if the user has set the global download option
       if (is_movie_check_enabled() )
@@ -71,6 +75,7 @@
   function category_add()
   {
     $cat     = rtrim(un_magic_quote($_REQUEST["cat_name"]));
+    $parent_id = $_REQUEST["parent_id"];
     $dl_info = rtrim(un_magic_quote($_REQUEST["dl_info"]));
     
     if(empty($cat))
@@ -83,10 +88,15 @@
         category_display('', '!'.str('CAT_ERROR_EXISTS'));
       else
       {
-        if(db_insert_row('categories', array('cat_name'=>$cat,'download_info'=>$dl_info)) === false)
+        if(db_insert_row('categories', array('cat_name'=>$cat,'parent_id'=>$parent_id,'download_info'=>$dl_info)) === false)
           category_display('', db_error());
         else
+        {
+          // Move media_locations from parent category
+          $cat_id = db_value("select cat_id from categories where cat_name='" . db_escape_str($cat) . "'");
+          db_sqlcommand("update media_locations set cat_id=$cat_id where cat_id=$parent_id");
           category_display('', str('CAT_ADDED_OK'));
+        }    
       }
     }
   }
@@ -109,17 +119,19 @@
     elseif(!empty($update_data))
     {
       $category_name = db_escape_str($update_data["CATEGORY"]);
+      $parent_id     = $update_data["CATEGORY_PARENT"];
       $download_info = db_escape_str($update_data["DOWNLOAD_INFO"]);
       $id = $update_data["CAT_IDS"];
-      
+        
       if(empty($category_name))
         category_display("!".str('CAT_ERROR_NAME'));
       else
       {
         db_sqlcommand("update categories
-                          set cat_name='$category_name' 
-                            , download_info='$download_info'
-                        where cat_id=$id");
+                       set cat_name='$category_name', parent_id=$parent_id, download_info='$download_info'
+                       where cat_id=$id");
+        // Move media_locations from parent category
+        db_sqlcommand("update media_locations set cat_id=$id where cat_id=$parent_id");
         category_display(str('CAT_UPDATE_OK'));
       }
     }
@@ -131,9 +143,17 @@
       {
         if($cat_id != 1)
         {
-          // Ensure that the existing media_locations are updated with no category
-          db_sqlcommand("update media_locations set cat_id=1 where cat_id=$cat_id");
-          db_sqlcommand("delete from categories where cat_id=$cat_id");
+          if (db_value("select count(*) from categories where parent_id=$cat_id")==0)
+          {
+            // Ensure that the existing media_locations are updated with no category
+            db_sqlcommand("update media_locations set cat_id=1 where cat_id=$cat_id");
+            db_sqlcommand("delete from categories where cat_id=$cat_id");
+          }
+          else
+          {
+            $category_name = db_value("select cat_name from categories where cat_id=$cat_id");
+            $message = "!".str('CAT_ERROR_DELETE',$category_name);
+          }
         }
         else
           $message = "!".str('CAT_ERROR_DELETE',$default_cat);
